requirements.txt
# Core Discord bot dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# Google YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Async and utility libraries
aiofiles>=23.0.0
asyncio-throttle>=1.0.2

# Data validation and parsing
validators>=0.20.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and monitoring (optional)
structlog>=23.0.0
